Refactored and added comments
