+++ src/nodes/models/TimeoutError.ts
@@ -1,0 +1,1 @@
+export class TimeoutError extends Error { }
